pyproject.toml
[project]
name = "django-ninja-template"
version = "0.1.0"
description = ""
authors = [
    {name = "gaojian",email = "olivetree123@163.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-ninja (>=1.4.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
